remove OmniBox.tsx and related tests, helpers, scss, etc. (moving the items still in use to GridPanel dir)
