+++ packages/components/src/public/QueryModel/FilterStatus.spec.tsx
@@ -3,10 +3,10 @@ import { mount, ReactWrapper } from 'enzyme';
 import { List } from 'immutable';
 import { Filter } from '@labkey/api';
 
-import { Value } from '../../internal/components/omnibox/Value';
-import { SearchAction } from '../../internal/components/omnibox/actions/Search';
-import { ViewAction } from '../../internal/components/omnibox/actions/View';
-import { FilterAction } from '../../internal/components/omnibox/actions/Filter';
+import { Value } from './grid/Value';
+import { SearchAction } from './grid/actions/Search';
+import { ViewAction } from './grid/actions/View';
+import { FilterAction } from './grid/actions/Filter';
 import { QueryInfo } from '../QueryInfo';
 
 import { FilterStatus } from './FilterStatus';
