+++ packages/components/src/public/QueryModel/SearchBox.tsx
@@ -1,7 +1,7 @@
 import React, { ChangeEvent, FC, FormEvent, memo, useCallback, useEffect, useState, useMemo } from 'react';
 
-import { ActionValue } from '../../internal/components/omnibox/actions/Action';
-import { filterActionValuesByType } from '../../internal/components/omnibox/utils';
+import { ActionValue } from './grid/actions/Action';
+import { filterActionValuesByType } from './grid/utils';
 
 interface Props {
     actionValues: ActionValue[];
