+++ packages/components/src/theme/index.scss
@@ -39,9 +39,7 @@
 @import "assays";
 @import "lineage";
 @import "report-list";
-@import "control";
 @import "menu";
-@import "multi";
 @import "navbar";
 @import "breadcrumb";
 @import "user";
