+++ packages/components/src/theme/menu.scss
@@ -1,112 +1,3 @@
-//
-// OmniBox Menu
-// ------------------------------
-
-
-// wrapper around the menu
-
-.OmniBox-menu-outer {
-  // Unfortunately, having both border-radius and allows scrolling using overflow defined on the same
-  // element forces the browser to repaint on scroll.  However, if these definitions are split into an
-  // outer and an inner element, the browser is able to optimize the scrolling behavior and does not
-  // have to repaint on scroll.
-  @include border-bottom-radius( $omnibox-input-border-radius );
-  background-color: $omnibox-input-bg;
-  border: 1px solid $omnibox-input-border-color;
-  border-top-color: mix($omnibox-input-bg, $omnibox-input-border-color, 50%);
-  box-shadow: 0 1px 0 rgba(0, 0, 0, 0.06);
-  box-sizing: border-box;
-  margin-top: -1px;
-  max-height: $omnibox-menu-max-height;
-  position: absolute;
-  top: 100%;
-  width: 100%;
-  z-index: $omnibox-menu-zindex;
-  -webkit-overflow-scrolling: touch;
-}
-
-
-// wrapper
-
-.OmniBox-menu {
-  max-height: ($omnibox-menu-max-height - 2px);
-  overflow-y: auto;
-}
-
-
-// options
-
-.OmniBox-option {
-  box-sizing: border-box;
-  background-color: $omnibox-option-bg;
-  color: $omnibox-option-color;
-  cursor: pointer;
-  display: block;
-  padding: $omnibox-padding-vertical $omnibox-padding-horizontal;
-
-  &:last-child {
-    @include border-bottom-radius( $omnibox-input-border-radius );
-  }
-
-  &.is-selected {
-    background-color: $omnibox-option-selected-bg;
-    color: $omnibox-option-selected-color;
-  }
-
-  &.is-focused {
-    background-color: $omnibox-option-focused-bg;
-    color: $omnibox-option-focused-color;
-  }
-
-  &.is-disabled {
-    color: $omnibox-option-disabled-color;
-    cursor: default;
-  }
-
-}
-
-// no results
-
-.OmniBox-noresults {
-  box-sizing: border-box;
-  color: $omnibox-noresults-color;
-  cursor: default;
-  display: block;
-  padding: $omnibox-padding-vertical $omnibox-padding-horizontal;
-}
-
-.OmniBox-autocomplete {
-  background-color: #fff;
-  min-width: 240px;
-  margin: 0;
-  z-index: 1000;
-  list-style: none;
-
-  border: 1px solid #e6e6e6;
-  padding: 0;
-  box-shadow: 0 2px 6px 0 rgba(0, 0, 0, 0.05);
-  border-radius: 3px;
-
-  position: absolute;
-
-  & > li {
-    height: 38px;
-    min-height: 38px;
-    max-height: 38px;
-    padding: 0 12px 0 20px;
-    display: flex;
-    align-items: center;
-    flex-direction: row;
-
-    cursor: pointer;
-    line-height: 1.4;
-    margin: 0;
-    border: 0;
-    outline: 0;
-    vertical-align: baseline;
-    list-style: none;
-  }
-}
 
 .dropdown-menu > li > div.dropdown-menu-row {
     display: block;
