--- packages/components/src/theme/multi.scss
@@ -1,119 +1,0 @@
-@import "variables";
-//
-// Multi-Select
-// ------------------------------
-
-
-// Base
-
-.OmniBox--multi {
-
-  // add margin to the input element
-  .OmniBox-input {
-    vertical-align: middle;
-    margin-left: $omnibox-padding-horizontal;
-    padding: 0;
-  }
-
-  // reduce margin once there is value
-  &.has-value .OmniBox-input {
-    margin-left: $omnibox-item-gutter;
-  }
-
-  // Items
-  .OmniBox-value {
-    white-space: pre;
-    display: inline-block;
-    position: relative;
-    border-radius: 3px;
-    margin: 5px 3px;
-    padding: 0 6px 0 30px;
-    background-color: $omnibox-item-bg;
-    border: 1px solid $omnibox-item-border-color;
-    cursor: pointer;
-    line-height: 22px;
-    max-width: calc(100% - 94px);
-    vertical-align: top;
-
-    i.read-lock {
-      padding-right: 6px;
-      vertical-align: middle;
-    }
-
-    i.symbol {
-      color: hsl(210,30%,50%);
-      margin-left: -25px;
-      padding-right: 6px;
-    }
-
-    &:hover,
-    &:focus {
-      background-color: $omnibox-item-hover-bg;
-    }
-    &:active {
-      background-color: $omnibox-item-border-color;
-    }
-
-    &.is-disabled {
-      background-color: $omnibox-item-disabled-bg;
-      border: 1px solid $omnibox-item-disabled-border-color;
-      color: $omnibox-item-disabled-color;
-
-      &:hover,
-      &:focus {
-        background-color: $omnibox-item-disabled-bg;
-      }
-      &:active {
-        background-color: $omnibox-item-disabled-bg;
-      }
-    }
-
-    &.is-readonly {
-      &:active {
-        background-color: $omnibox-item-hover-bg;
-      }
-    }
-  }
-
-  // common
-  .OmniBox-value-icon,
-  .OmniBox-value-label {
-    display: inline-block;
-    vertical-align: middle;
-  }
-
-  // label
-  .OmniBox-value-label {
-    @include border-right-radius( $omnibox-item-border-radius );
-    cursor: default;
-    padding: $omnibox-item-padding-vertical $omnibox-item-padding-horizontal;
-  }
-  a.OmniBox-value-label {
-    color: $omnibox-item-color;
-    cursor: pointer;
-    text-decoration: none;
-
-    &:hover {
-      text-decoration: underline;
-    }
-  }
-
-  // icon
-  .OmniBox-value-icon {
-    cursor: pointer;
-    @include border-left-radius( $omnibox-item-border-radius );
-    border-right: 1px solid $omnibox-item-border-color;
-
-    // move the baseline up by 1px
-    padding: ($omnibox-item-padding-vertical - 1) $omnibox-item-padding-horizontal ($omnibox-item-padding-vertical + 1);
-
-    &:hover,
-    &:focus {
-      background-color: $omnibox-item-hover-bg;
-      color: $omnibox-item-hover-color;
-    }
-    &:active {
-      background-color: $omnibox-item-border-color;
-    }
-  }
-}
